requirements.txt
streamlit
openai-whisper
ollama
torch
torchaudio
python-dotenv
numpy
scipy
requests
pydub
librosa
soundfile
